requirements.txt
langchain
langchain-classic
langchain-community
langchain-openai
langchain-text-splitters
langchain-huggingface
langchain_groq
langchain_core
langserve

chromadb
langchain-chroma
faiss-cpu
sentence-transformers

pypdf
pymupdf
arxiv
wikipedia
beautifulsoup4
bs4

python-dotenv
ipykernel
streamlit
fastapi
uvicorn
sse_starlette
ddgs

